# docs/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API de Gestión de Tareas",
    "description": "API REST para gestionar tareas pendientes",
    "contact": {
      "name": "Mariano Bonansea",
      "email": "marianobonanseapetrovial@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Generated server url" }
  ],
  "tags": [{ "name": "Tareas", "description": "API para gestionar tareas" }],
  "paths": {
    "/api/tasks": {
      "get": {
        "tags": ["Tareas"],
        "summary": "Obtener todas las tareas",
        "operationId": "getTasks",
        "responses": {
          "200": {
            "description": "Lista de tareas recuperada exitosamente",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tareas"],
        "summary": "Crear una nueva tarea",
        "operationId": "createTask",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "string" } } },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Tarea creada exitosamente",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/Task" } }
            }
          }
        }
      }
    },
    "/api/tasks/{id}/toggle": {
      "patch": {
        "tags": ["Tareas"],
        "summary": "Cambiar estado de una tarea",
        "operationId": "toggleTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/Task" } }
            }
          }
        }
      }
    },
    "/api/tasks/{id}": {
      "delete": {
        "tags": ["Tareas"],
        "summary": "Eliminar una tarea",
        "operationId": "deleteTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "completed": { "type": "boolean" }
        }
      }
    }
  }
}